udp: extract apple_fast module
